// src/types.ts
export interface User {
  id: string;
  email: string;
  role: "Administrator" | "Creator" | "Reviewer" | "Endorser" | "Contributor";
  full_name: string;
  position: string;
  deptId?: string;
  programmeId?: string;
}

export interface Session {
  id: string;
  name: string;
  isActive: boolean;
  isArchived: boolean;
}

export interface Department {
  id: string;
  name: string;
  code: string;
}

export interface Programme {
  id: string;
  deptId: string;
  name: string;
  code: string;
  headOfProgramme?: string;
}